+++ backend/coreapp/compiler_wrapper.py
@@ -1,4 +1,6 @@
+from typing import Optional
 from coreapp.models import Asm, Assembly, Compilation
+from coreapp import util
 from django.conf import settings
 from django.utils.crypto import get_random_string
 import json
@@ -31,6 +33,16 @@ def load_compilers() -> dict:
 
 compilers = load_compilers()
 
+
+def check_compilation_cache(*args) -> Optional[Compilation]:
+    hash = util.gen_hash(args)
+    return Compilation.objects.filter(hash=hash).first()
+
+def check_assembly_cache(*args) -> Optional[Compilation]:
+    hash = util.gen_hash(args)
+    return Assembly.objects.filter(hash=hash).first()
+
+
 class CompilerWrapper:
     def base_path():
         return settings.COMPILER_BASE_PATH
@@ -45,13 +57,13 @@ class CompilerWrapper:
             .replace("$CC_OPTS", cc_opts) \
             .replace("$AS_OPTS", as_opts)
         compile_command = f"bash -c \"{compile_command}\""
-        # TODO sandbox
 
         logger.debug(f"Compiling: {compile_command}")
 
         return_code = 0
 
         try:
+            # TODO sandbox
             result = subprocess.run(compile_command, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
             stderr = result.stderr.decode()
 
@@ -106,6 +118,7 @@ class CompilerWrapper:
         logger.debug(f"Objdumping: {objdump_command}")
 
         try:
+            # TODO sandbox
             result = subprocess.run(objdump_command, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
 
             if result.returncode != 0:
@@ -121,11 +134,18 @@ class CompilerWrapper:
 
     @staticmethod
     def compile_code(compiler: str, cpp_opts: str, as_opts: str, cc_opts: str, code: str, context: str):
-        compiler_path = CompilerWrapper.base_path() / compiler.shortname
+        compiler_path = CompilerWrapper.base_path() / compiler
+
+        compile_cfg = compilers[compiler]
 
         if not compiler_path.exists():
-            logger.error(f"Compiler {compiler.shortname} not found")
+            logger.error(f"Compiler {compiler} not found")
             return (None, "ERROR: Compiler not found")
+
+        cached_compilation = check_compilation_cache(compiler, cpp_opts, as_opts, cc_opts, code, context)
+        if cached_compilation:
+            logger.debug(f"Compilation cache hit!")
+            return (cached_compilation, cached_compilation.stderr)
 
         temp_name = get_random_string(length=8)
         settings.COMPILATION_OBJECTS_PATH.mkdir(exist_ok=True, parents=True)
@@ -143,7 +163,7 @@ class CompilerWrapper:
 
         # Run compiler
         compile_status, stderr = CompilerWrapper.run_compiler(
-            compiler.compile_cmd,
+            compile_cfg["cc"],
             code_path,
             object_path,
             compiler_path,
@@ -161,19 +181,28 @@ class CompilerWrapper:
             return (None, stderr)
         
         # Store Compilation to db
-        compilation = Compilation(compiler, cpp_opts, as_opts, cc_opts, code, context, object_path)
+        compilation = Compilation(compiler, cpp_opts, as_opts, cc_opts, code, context, object_path, stderr)
         compilation.save()
 
         return (compilation, stderr)
 
     @staticmethod
     def assemble_asm(compiler:str, as_opts: str, asm: Asm, to_overwrite:Assembly = None) -> Assembly:
-        compiler_path = CompilerWrapper.base_path() / compiler.shortname
+        compiler_path = CompilerWrapper.base_path() / compiler
 
         if not compiler_path.exists():
-            logger.error(f"Compiler {compiler.shortname} not found")
+            logger.error(f"Compiler {compiler} not found")
             return "ERROR: Compiler not found"
         
+        # Check the cache if we're not manually re-running an Assembly
+        if not to_overwrite:
+            cached_assembly = check_assembly_cache(compiler, as_opts, asm)
+            if cached_assembly:
+                logger.debug(f"Assembly cache hit!")
+                return cached_assembly
+
+        compiler_cfg = compilers[compiler]
+
         assemblies_path = settings.ASM_OBJECTS_PATH
         assemblies_path.mkdir(exist_ok=True, parents=True)
         
@@ -186,7 +215,7 @@ class CompilerWrapper:
             f.write(ASM_MACROS + asm.data)
 
         assemble_status, stderr = CompilerWrapper.run_assembler(
-            compiler.assemble_cmd,
+            compiler_cfg["as"],
             asm_path,
             object_path,
             compiler_path,
@@ -200,13 +229,12 @@ class CompilerWrapper:
             if object_path.exists():
                 os.remove(object_path)
             return None #f"ERROR: {assemble_status[1]}"
-        
-        # Store Assembly to db
-        assembly = Assembly(compiler, as_opts, source_asm=asm, object=object_path)
 
         if to_overwrite:
             assembly = to_overwrite
             assembly.object = object_path
+        else:
+            assembly = Assembly(compiler, as_opts, source_asm=asm, object=object_path)
 
         assembly.save()