+++ backend/coreapp/urls.py
@@ -3,8 +3,8 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path('', views.index, name='index'),
     path('scratch', views.scratch, name='scratch'),
     path('scratch/<slug:slug>', views.scratch, name='scratch'),
     path('scratch/<slug:slug>/compile', views.compile, name='compile_scratch'),
+    path('scratch/<slug:slug>/fork', views.fork, name='fork_scratch'),
 ]