+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-# 0.2
+# 8/8/2021
 
 * Improved performance by only sending context to the backend if it was changed from the current version
 
+* Added scratch forking